feat: CA-261 - patch `cursor` directly on `BaseDatabaseWrapper`
this used to be done this way until a7541a1, when it was changed to the
`.im_class` method apparently for compatibility with Django Debug
Toolbar. Python 3 does not have `.im_class` and unbound methods are
plain functions, so it is no longer possible to determine the class
where the method is actually defined. Since we no longer use Django
Debug Toolbar but we do want Python 3, we're going back to the old way
of patching